[Chore #154082903] Reorganise file structure for more intuitive arrangement of component source code.
