--- src/components/app.js
@@ -1,14 +1,0 @@
-import React, { Component } from "react";
-import SearchBar from "../containers/search_bar";
-import SearchResult from "../containers/search_result";
-
-export default class App extends Component {
-  render() {
-    return (
-      <div>
-        <SearchBar />
-        <SearchResult />
-      </div>
-    );
-  }
-}
