+++ src/components/auth/create_account.js
@@ -1,15 +1,15 @@
 import React, { Component } from "react";
 import { Field, reduxForm } from "redux-form";
 import { Link } from "react-router-dom";
-import { changePassword } from "../actions/index";
+import { createUser } from "../../actions/index";
 import { connect } from "react-redux";
-import NavBarUnAuthed from "./nav_bar_unauthed";
+import NavBarUnAuthed from "../nav/guest";
 
-class ChangePassword extends Component {
+class CreateAccount extends Component {
   onSubmit(values) {
     console.log("Form has been submited:", values);
-    this.props.changePassword(values, () => {
-      this.props.history.push("/shoppinglists");
+    this.props.createUser(values, () => {
+      this.props.history.push("/");
     });
   }
 
@@ -31,28 +31,35 @@ class ChangePassword extends Component {
     const { handleSubmit } = this.props;
     return (
       <div>
+        <NavBarUnAuthed />
         <div className="row">
           <div className="col-md-5 col-md-offset-4 well well-lg">
             <div id="login">
-              <h1>Change Password</h1>
+              <h1>Register for an account</h1>
               <form
                 className="form-horizontal"
                 onSubmit={handleSubmit(this.onSubmit.bind(this))}
               >
                 <Field
+                  name="email"
+                  label="Email"
+                  type="text"
+                  component={this.renderField}
+                />
+                <Field
                   name="password"
-                  label="New Password"
+                  label="Password"
                   type="password"
                   component={this.renderField}
                 />
                 <Field
                   name="repassword"
-                  label="Confirm New Password"
+                  label="Confirm Password"
                   type="password"
                   component={this.renderField}
                 />
                 <button type="submit" className="btn btn-primary">
-                  Change{" "}
+                  Register{" "}
                 </button>
                 <Link to="/" className="btn btn-danger">
                   Cancel
@@ -68,6 +75,9 @@ class ChangePassword extends Component {
 
 function validate(values) {
   const errors = {};
+  if (!values.email || values.email.length < 3) {
+    errors.email = "Enter title that is more than three characters !";
+  }
 
   if (!values.password || values.password.length < 3) {
     errors.password = "Enter title that is more than three characters !";
@@ -80,5 +90,5 @@ function validate(values) {
 
 export default reduxForm({
   validate,
-  form: "ChangePasswordForm"
-})(connect(null, { changePassword })(ChangePassword));
+  form: "CreateAccountForm"
+})(connect(null, { createUser })(CreateAccount));
