+++ src/components/misc/pagination.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import { fetchLists } from "../actions/index";
+import { fetchLists } from "../../actions/index";
 import { Link } from "react-router-dom";
 
 class Pagination extends Component {
