+++ src/components/search/search_result.js
